requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
audioop-lts==0.2.1
babel==2.16.0
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.11.0
blivet-gui==2.6.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cobble==0.1.4
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
defusedxml==0.7.1
distro==1.9.0
dnf==4.22.0
et_xmlfile==2.0.0
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.9.0
idna==3.7
iso639==0.1.4
jiter==0.8.2
langtable==0.0.68
libcomps==0.1.21
libdnf==0.73.4
libretranslatepy==2.1.1
louis==3.28.0
lxml==5.2.1
Mako==1.2.3
mammoth==1.9.0
markdownify==0.14.1
markitdown==0.0.1a3
MarkupSafe==2.1.5
msgpack==1.0.6
mutagen==1.47.0
nftables==0.1
numpy==2.2.2
olefile==0.47
openai==1.60.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
Paste==3.10.1
pathvalidate==3.2.3
pdfminer.six==20240706
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
ply==3.11
productmd==1.43
progressbar2==3.53.2
ptyprocess==0.7.0
puremagic==1.28
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.21.0
pycups==2.0.4
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
pyenchant==3.2.2
pygit2==1.17.0
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pynvim==0.5.2
pyOpenSSL==24.2.1
pyparted==3.13.0
PyPDF2==3.0.1
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-pptx==1.0.2
python-utils==3.7.0
pytz==2024.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.20.0
rpmautospec==0.7.3
rpmautospec-core==0.1.5
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.8.2
soupsieve==2.6
SpeechRecognition==3.14.0
standard-aifc==3.13.0
standard-chunk==3.13.0
systemd-python==235
Tempita==0.5.2
tqdm==4.67.1
translate==3.6.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
websockets==12.0
xkbregistry==0.3
XlsxWriter==3.2.0
youtube-transcript-api==0.6.3
yt-dlp==2025.1.26
